// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/employeeratetagconfiguration.ts
export class EmployeeRateTagConfigurationModel {
    rateTagEmployeeId: string;
    startDate: Date;
    endDate: Date;
    isInHerited:boolean; 
    rateTagRoleBranchConfigurationId: string;
    branchId: string;
    roleId: string;
    employeeName: string;
    branchName: string;
    roleName: string;
    priority: number;
    employeeId: string;
}