// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/softlabels-model.ts
export class SoftLabelConfigurationModel {
    softLabelConfigurationId: string;
    projectLabel: string;
    goalLabel: string;
    employeeLabel: string;
    userStoryLabel: string;
    deadlineLabel: string;
    projectsLabel: string;
    goalsLabel: string;
    employeesLabel: string;
    userStoriesLabel: string;
    deadlinesLabel: string;
    estimatedTimeLabel: string;
    estimationLabel: string;
    estimationsLabel: string;
    estimateLabel: string;
    estimatesLabel: string;
    companyId: string;
    createdDateTime: string;
    timeStamp: any;
    searchText: string;
}